% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{calc_corr}
\alias{calc_corr}
\title{Calculate Correlation of Pixel Values between Two Channels}
\usage{
calc_corr(img1, img2, pix_ss)
}
\arguments{
\item{img1}{Array 3d: tiles of the 1st channel for which correlation is sought with the 2nd}

\item{img2}{Array 3d: tiles of the 2nd channel for which correlation is sought with the 1st}

\item{pix_ss}{Integer: number of pixels to randomly subsample for calculating correlation}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Numeric: absolute value of correlation between the two channels
}\if{html}{\out{</div>}}
}
\description{
Calculate correlation of pixel values between two channels.
}
\details{
This function takes two channels, already restricted to just the correct subset of tiles, and
calculates the pixel-wise correlation between the two channels. If pixels will be randomly subsampled,
this is done first. The absolute value of the correlation is returned.
}
