% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{calc_MI}
\alias{calc_MI}
\title{Calculate Mutual Information of Pixel Values between Two Channels}
\usage{
calc_MI(img1, img2, pix_ss)
}
\arguments{
\item{img1}{Array 3d: tiles of the 1st channel for which MI is sought with the 2nd}

\item{img2}{Array 3d: tiles of the 2nd channel for which MI is sought with the 1st}

\item{pix_ss}{Integer: number of pixels to randomly subsample for calculating MI}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Numeric: normalized mutual information between the two channels
}\if{html}{\out{</div>}}
}
\description{
Calculate mutual information of pixel values between two channels.
}
\details{
This function takes two channels, already restricted to just the correct subset of tiles, and
calculates the pixel-wise mutual information between the two channels. If pixels will be randomly subsampled,
this is done first. The normalized mutual information is returned.
}
