% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{calc_Pmat}
\alias{calc_Pmat}
\title{Assemble the Correction Matrix}
\usage{
calc_Pmat(img, method, q_thr, tile_size, stride, pix_ss, cores)
}
\arguments{
\item{img}{Array 3d: all channels of an image}

\item{method}{Character: whether to minimize mutual info "MI" or correlation "Cor"}

\item{q_thr}{Numeric: quantile threshold to select only the tiles with lowest SSIM}

\item{tile_size}{Integer: dimensions of each square tile in pixels}

\item{stride}{Integer: distance between top left corner of each tile in pixels}

\item{pix_ss}{Integer: number of pixels to randomly subsample for calculating MI or correlation}

\item{cores}{Integer: number of cores to use for parallel calculations}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Matrix: the fully calculated correction matrix
}\if{html}{\out{</div>}}
}
\description{
Assemble the correction matrix
}
\details{
This function orchestrates the inference of the correction scaling factor for every pair of channels.
Entry i,j in the correction matrix gives the scaling factor to apply to image i before subtracting it
from image j, in order to correct image j based on overlap with image i.
}
