% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{calc_ssim}
\alias{calc_ssim}
\title{Calculate SSIM between Tiles of Two Channels}
\usage{
calc_ssim(tiles)
}
\arguments{
\item{tiles}{List: 3d arrays of tiles for two channels}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Vector: SSIM between corresponding pairs of tiles
}\if{html}{\out{</div>}}
}
\description{
Calculate SSIM between tiles of two channels
}
\details{
This function calculates the Structural Similarity Index Measure, SSIM, between corresponding tiles
across two channels of an image.
}
