% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{chop_image}
\alias{chop_image}
\title{Convert an Image into Tiles}
\usage{
chop_image(img, tile_size, stride)
}
\arguments{
\item{img}{Array 3d: all channels of an image}

\item{tile_size}{Integer: dimensions of each square tile in pixels}

\item{stride}{Integer: distance between top left corner of each tile in pixels}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ List: for each channel of the image, a 3d array of the tiles
}\if{html}{\out{</div>}}
}
\description{
Convert an image into tiles
}
\details{
This function cuts an image into tiles of specified size and spacing. Each channel becomes a 3d array,
where x and y correspond to x and y in the original image, and z is the tile ID. The 3d arrays representing
each channel are returned in a list.
}
