% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{load_image}
\alias{load_image}
\title{Load an Image}
\usage{
load_image(file)
}
\arguments{
\item{file}{Character: path to 8-bit grayscale multichannel .tif image}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Array 3d: a 3d array representation of the original image
}\if{html}{\out{</div>}}
}
\description{
Load an image
}
\details{
This function loads an 8-bit grayscale multichannel .tif image and transforms it into a 3d array.
}
