% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{mosaic_picasso}
\alias{mosaic_picasso}
\title{Apply Mosaic-PICASSO Correction to an Image}
\usage{
mosaic_picasso(
  file,
  method = "MI",
  q_thr = 0.5,
  tile_size = 50,
  stride = 50,
  pix_ss = "all",
  enhance = T,
  groups = NULL,
  cycles = 1,
  cores = NULL
)
}
\arguments{
\item{file}{Character: path to 8-bit grayscale multichannel .tif image}

\item{method}{Character: whether to minimize mutual info "MI" or correlation "Cor"}

\item{q_thr}{Numeric: quantile threshold to select only the tiles with lowest SSIM}

\item{tile_size}{Integer: dimensions of each square tile in pixels}

\item{stride}{Integer: distance between top left corner of each tile in pixels}

\item{pix_ss}{Integer: number of pixels to randomly subsample for calculating MI or correlation}

\item{enhance}{Logical: whether to enhance contrast in the corrected image}

\item{groups}{List: vectors of channel numbers to include in self-contained groups for correction}

\item{cycles}{Integer: number of iterations of correction to perform}

\item{cores}{Integer: number of cores to use for parallel calculations}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Nothing: a corrected .tif and a .csv of the correction matrix are saved to the working directory
}\if{html}{\out{</div>}}
}
\description{
Apply Mosaic-PICASSO Correction to an Image
}
\details{
This function orchestrates the entire Mosaic-PICASSO algorithm to infer and remove autofluorescence
and spillover from a multiplex image. Steps include loading the image, splitting the image into tiles,
selecting only tiles with the lowest structural similarity index measure, calculating the matrix of scaling
factors for correction, applying the correction to each channel, optionally enhancing the contrast of
the corrected image, and saving the output. The input image must be an 8-bit grayscale multichannel .tif.
The output is an image of the same format, with "_NEW" appended to its name, along with a .csv copy of
the correction matrix.
}
