% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{update_Pmat}
\alias{update_Pmat}
\title{Update a Correction Matrix}
\usage{
update_Pmat(p_mat_0, p_mat_1)
}
\arguments{
\item{p_mat_0}{Matrix: initial correction matrix}

\item{p_mat_1}{Matrix: new correction matrix}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Matrix: updated correction matrix
}\if{html}{\out{</div>}}
}
\description{
Update a correction matrix by multiplying an original correction matrix with a new one.
}
\details{
This function multiplies an initial correction matrix by a second one. This is only meaningful if
more than one cycle of correction is being performed.
}
